Add files via upload
This is a repository for my research project titled, "An Innovative Approach To Tuberculosis Detection/Diagnosis Using Machine Learning/Deep Learning ".  In this project, I explored the application of deep learning models, specifically VGG16, ResNet50, Inceptionv3, and DenseNet121, for the detection of tuberculosis (TB) using chest X-ray images. Through the implementation of transfer learning techniques and extensive data preprocessing, including the handling of class imbalance, each model was trained and tested to assess its effectiveness in accurately identifying TB. Among the models evaluated, DenseNet121 demonstrated superior performance, achieving an impressive accuracy of 98% on the test dataset, outperforming the other models.
The findings underscore the potential of DenseNet121 as a highly effective tool for TB detection, offering a promising approach to improving the speed and accuracy of TB diagnosis.